+++ packages/expo-media-library/src/next/index.ts
@@ -14,22 +14,6 @@ export class Query extends ExpoMediaLibraryNext.Query {}
 
 export class Asset extends ExpoMediaLibraryNext.Asset {
   // @hidden
-  getFavorite(): Promise<boolean> {
-    if (Platform.OS !== 'ios') {
-      throw new UnavailabilityError('MediaLibrary', 'getFavorite is only available on iOS');
-    }
-    return super.getFavorite();
-  }
-
-  // @hidden
-  setFavorite(isFavorite: boolean): Promise<void> {
-    if (Platform.OS !== 'ios') {
-      throw new UnavailabilityError('MediaLibrary', 'setFavorite is only available on iOS');
-    }
-    return super.setFavorite(isFavorite);
-  }
-
-  // @hidden
   getMediaSubtypes(): Promise<MediaSubtype[]> {
     if (Platform.OS !== 'ios') {
       throw new UnavailabilityError('MediaLibrary', 'getMediaSubtypes is only available on iOS');